Jun, 01, 2018. ArtPDE project revised
