+++ CMakeLists.txt
@@ -8,6 +8,7 @@ set(CMAKE_CXX_STANDARD 11)
 
 include_directories(${CMAKE_SOURCE_DIR})
 
+file(GLOB GeometrySource "Project/*.hpp")
 file(GLOB GeometrySource "BasicData/*.hpp")
 file(GLOB GeometrySource "Geometry/*.hpp")
 #file(GLOB GeometrySource "Geometry_New/*.hpp")
