+++ OldCode/GeometricData/edge.hpp
