+++ OldCode/GeometricData/vertex.hpp
