+++ main_kavy.cxx
@@ -1,12 +1,16 @@
 #include "Demo/DemoPositionVector.hpp"
 #include "Demo/DemoGeometry.hpp"
+#include "Demo/DemoArtProject.hpp"
 
 using namespace art_pde;
 
 int main() {
 //    DemoPositionVector();
 
-    DemoGeometry();
+    DemoArtProject();
+
+
+//    DemoGeometry();
 
 
     return 0;
